% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxGCAR-package.R
\docType{package}
\name{proxGCAR-package}
\alias{proxGCAR}
\alias{proxGCAR-package}
\title{proxGCAR: Gaussian Conditional Autoregressive MLE}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
