% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaximumLikelihood.R
\name{Maximum_Likelihood}
\alias{Maximum_Likelihood}
\title{Computes the Maximum Likelihood Estimators for a Gaussian CAR model and their variances}
\usage{
Maximum_Likelihood(Y, proximity)
}
\arguments{
\item{Y}{A nx1 vector that represents an singular observation of}

\item{proximity}{A proximity matrix}
}
\value{
A matrix of LS_estimators , MLE's , and MLE variances for rho(p) , tau , mu
}
\description{
Computes the Maximum Likelihood Estimators for a Gaussian CAR model and their variances
}
